fix: store wishlist and cart when not logged in

+++ src/context/index.tsx
@@ -5,12 +5,11 @@ import { createContext } from 'react';
 import { Session, createClient, SupabaseClient } from '@supabase/supabase-js';
 import { Database } from '../../database.types'
 import { toast } from 'react-toastify';
-import { genErrorMessage } from '../utils';
+import { delCart, deleteWishilist, genErrorMessage, updateWishlist, upsertCart } from '../utils';
 
 const supabaseURL = import.meta.env.VITE_SUPABASE_URL;
 const supabaseAnonKey = import.meta.env.VITE_SUPABASE_ANON_KEY;
 
-console.log(supabaseAnonKey, supabaseURL)
 
 interface Auth {
     session: Session | null,
@@ -20,8 +19,8 @@ interface Auth {
     addToWishlist: (id: string) => Promise<number>,
     removeFromWishlist: (id: string) => Promise<number>,
     fetchCart: () => Promise<any[] | null>,
-    addToCart: (id: string, size: { label: string, id: string }, gender: string) => Promise<number>,
-    removeFromCart: (id: string) => Promise<number>,
+    addToCart: (id: string, size: string, gender: string) => Promise<number>,
+    removeFromCart: (id: string, force?: boolean) => Promise<number>,
     wishlist: any[],
     cart: any[],
 }
@@ -58,14 +57,15 @@ const DBContext = createContext<Auth>({
       });
       return response;
     },
-    addToCart: async (id: string, size: { label: string, id: string }, gender: string) => {
+    addToCart: async (id: string, size: string, gender: string) => {
       const response = await new Promise<number>((resolve) => {
         console.log(id, size, gender);
         resolve(200);
       });
       return response;
     },
-    removeFromCart: async (id: string) => {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    removeFromCart: async (id: string, _force: boolean = false) => {
       const response = await new Promise<number>((resolve) => {
         console.log(id);
         resolve(200);
@@ -98,7 +98,7 @@ const DBProvider = ({ children }: { children: React.ReactNode }) => {
       if (!supabaseClient) return;
       fetchWishlist();
       fetchCart();
-    }, [supabaseClient])
+    }, [supabaseClient, session])
 
    
     const signOut = async () => {
@@ -121,6 +121,11 @@ const DBProvider = ({ children }: { children: React.ReactNode }) => {
 
     const fetchWishlist = async () => {
       if (!supabaseClient) return null;
+      if (!session) {
+        const data = JSON.parse(localStorage.getItem('wishlistData') ?? '[]')
+        setWishlist(data);
+        return data;
+      }
       const { data, error } = await supabaseClient.from('wishlist').select();
       setWishlist(data as any);
       if (error) {
@@ -132,32 +137,45 @@ const DBProvider = ({ children }: { children: React.ReactNode }) => {
 
     const addToWishlist = async (product_id: string) => {
       if (!supabaseClient) return 404;
-      const { error, status } = await supabaseClient
+      let s = 200;
+      if (session) {
+        const { error, status } = await supabaseClient
         .from('wishlist')
         .insert({ product_id })
-      setWishlist((w)=> [...w, { product_id }])
-      if (error) {
-        toast(genErrorMessage(error.code, 'wishlist'), { type: 'error' })
+        if (error) {
+          toast(genErrorMessage(error.code, 'wishlist'), { type: 'error' })
+        }
+        s = status;
       }
-      return status;
+      setWishlist((w)=> updateWishlist(w, { product_id }))
+      return s;
     }
 
     const removeFromWishlist = async (product_id: string) => {
       if (!supabaseClient) return 404;
-      const { error, status } = await supabaseClient
-        .from('wishlist')
-        .delete()
-        .eq('product_id', product_id)
-      setWishlist((w)=> w.filter(x => x['product_id'] !== product_id));
-      if (error) {
-        toast(genErrorMessage(error.code, 'wishlist'), { type: 'error' })
+      let s = 200;
+      if (session) {
+        const { error, status } = await supabaseClient
+          .from('wishlist')
+          .delete()
+          .eq('product_id', product_id)
+        if (error) {
+          toast(genErrorMessage(error.code, 'wishlist'), { type: 'error' })
+        }
+        s = status;
       }
-      return status;
+      setWishlist((w)=> deleteWishilist(w, { product_id }));
+      return s;
     }  
 
     const fetchCart = async () => {
       if (!supabaseClient) return null;
-      const { data, error } = await supabaseClient.from('wishlist').select();
+      if (!session) {
+        const data = JSON.parse(localStorage.getItem('cartData') ?? '[]')
+        setCart(data);
+        return data;
+      }
+      const { data, error } = await supabaseClient.from('cart').select();
       setCart(data as any);
       if (error) {
         toast(genErrorMessage(error.code, 'cart'), { type: 'error' })
@@ -166,36 +184,50 @@ const DBProvider = ({ children }: { children: React.ReactNode }) => {
       return data
     }
 
-    const addToCart = async (product_id: string, size: { label: string, id: string }, gender: string) => {
-      if (!supabaseClient) return 404;
-      const { error, status } = await supabaseClient
+    const addToCart = async (product_id: string, size: string, gender: string) => {
+      let s: number = 200;
+      const p_id = `${product_id}-${size}-${gender}`;
+      if (session) {
+        if (!supabaseClient) return 404;
+        const { error, status } = await supabaseClient
           .rpc('upsert_to_cart', {
             p_gender: gender, 
-            p_id: `${product_id}-${size.id}-${gender}`, 
+            p_id,
             p_product_id: product_id, 
             p_quantity: 1,
-            p_size: size.label,
+            p_size: size,
           });
-    
-      // setCart((c)=> [...c, { product_id, size, gender }])
+          if (error) {
+            toast(genErrorMessage(error.code, 'cart'), { type: 'error' })
+          }
+          s = status;
+      }
+     
+      setCart((c)=> [...upsertCart(c,{ id: p_id, product_id, quantity: 1, size, gender })])
+      return s;
+    }
+
+    const removeFromCart = async (unique_id: string, force: boolean = false) => {
+      if (!supabaseClient) return 404;
+      if (force) {
+        const { error, status } = await supabaseClient.from('cart').delete().eq('id', unique_id);
+        if (error) {
+          toast(genErrorMessage(error.code, 'cart'), { type: 'error' })
+        }
+        setCart((c) => [...c.filter(x => x.id !== unique_id)]);
+        return status;
+      }
+      const { error, status } = await supabaseClient
+          .rpc('delete_from_cart', {
+            p_id: unique_id, 
+          });
+      setCart((c) => [...delCart(c, unique_id)]);
       if (error) {
         toast(genErrorMessage(error.code, 'cart'), { type: 'error' })
       }
       return status;
     }
 
-    const removeFromCart = async (unique_id: string) => {
-      if (!supabaseClient) return 404;
-      const { error, status } = await supabaseClient
-          .rpc('delete_from_cart', {
-            p_id: unique_id, 
-          });
-      // setCart((w)=> w.filter(x => x['product_id'] !== product_id));
-      if (error) {
-        toast(genErrorMessage(error.code, 'cart'), { type: 'error' })
-      }
-      return status;
-    }  
 
     return <DBContext.Provider value={{
       cart,
